+++ lib/bunko/models/post_type_methods.rb
@@ -10,8 +10,8 @@ module Bunko
         has_many :posts, dependent: :restrict_with_error
 
         # Validations
-        validates :name, presence: true
-        validates :slug, presence: true, uniqueness: true
+        validates :name, presence: true, uniqueness: true
+        validates :title, presence: true
       end
     end
   end